requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
selenium==4.15.2
pillow==10.1.0
python-multipart==0.0.6
pydantic==2.5.2
